// src/enums/filenames.ts
export enum Filenames {
  ESLINT_CONFIG = '.eslintrc.js',
  PACKAGE_JSON = 'package.json',
  NODEMON_CONFIG = 'nodemon.json',
  TS_CONFIG = 'tsconfig.json',
  PRETTIER_CONFIG = '.prettierrc.js',
  JEST_TEST = 'hello.spec.ts',
  JEST_CONFIG = 'jest.config.js',
  SONARQUBE_PROPERTIES = 'sonar-project.properties',
  SONARQUBE_REPORT_SCRIPT = 'report_sonarqube.sh',
  DOCKER_COMPOSE = 'docker-compose.yml',
  GIT_IGNORE = '.gitignore',

  SRC_FOLDER = 'src',
  SRIPTS_FOLDER = 'scripts',
}